hpb: make ExtensionRegistry backend-aware
When the backend is upb, the registry_ var will be initted and supplied an arena. One can also add an extension, and get the generated registry (under the hood, uses upb linker arrays).

The cpp backend doesn't really have a notion of an ExtensionRegistry, so this effectively reduces down to a thin no-op wrapper. Since the C++ backend can't construct an ExtensionRegistry directly, hpb functions must be called with hpb::ExtensionRegistry::EmptyRegistry().

PiperOrigin-RevId: 784215543